backend/requirements.txt
# general
Django==4.2.1
Pillow==9.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2

# clean code
flake8==6.0.0
